requirements.txt
aiohttp~=3.8.1
black==22.6.0
pytest==7.1.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-env==0.6.2
pyparsing==3.0.9
aiohttp-cors==0.7.0
aiohttp-apispec==2.2.3
aiohttp-session==2.11.0
aioresponses==0.7.3
alembic==1.8.1
marshmallow==3.19.0
SQLAlchemy==1.4.40
apispec==3.3.2
async-timeout==4.0.2
coverage==6.4.4
cryptography==37.0.4

PyYAML~=6.0